Render the Moon using a model with altitude details
The model is called "striped" to reflect the fact that it has high
detalization on the sides of the lunar disk as seen from the Earth.
This lets us have correct silhouette of the Moon e.g. when it covers
the Sun, or when we test for occlusion of stars.
To have the correct silhouette when seen from any other object we must
have a more complete model, which would take 52 MiB (while the current
one is 9.6 MiB).

The model is not in Wavefront OBJ format, because the textual nature
would inflate the size of already rather large model and reduce loading
speed. Instead, the format is as follows:

0     uint32        number of vertices, N
4     uint32        number of indices, M
8     float32[3*N]  array of vertices
8+3*N uint32[M]     array of indices