Re-enable antialiasing for panel borders
It was broken by the switch to a non-multisampled final framebuffer, so
now the implementation uses a QPixmap rendered with QPainter.

Oculars panel is rendered completely as a single QGraphicsPixmapItem,
because it's a rather static rectangle.

Left and bottom panels are made of a path that defines the background, a
path that defines straight lines, and two pixmaps for the curvy borders
in the corners. This is done to avoid redrawing the bounding rectangle
of these two panels, which is rather large and would need an update on
every frame when the panels' motion is animated.